Replace auth.js file #4
